refactor(geo): simplify adapters and isolate boundary tests
